+++ docs/code/interfaces/types_debug_utils.PersistSourcemapsParams.md
@@ -1,0 +1,54 @@
+[@algorandfoundation/algokit-utils](../README.md) / [types/debug-utils](../modules/types_debug_utils.md) / PersistSourcemapsParams
+
+# Interface: PersistSourcemapsParams
+
+[types/debug-utils](../modules/types_debug_utils.md).PersistSourcemapsParams
+
+## Table of contents
+
+### Properties
+
+- [client](types_debug_utils.PersistSourcemapsParams.md#client)
+- [projectRoot](types_debug_utils.PersistSourcemapsParams.md#projectroot)
+- [sources](types_debug_utils.PersistSourcemapsParams.md#sources)
+- [withSources](types_debug_utils.PersistSourcemapsParams.md#withsources)
+
+## Properties
+
+### client
+
+• **client**: `default`
+
+#### Defined in
+
+[src/types/debug-utils.ts:88](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L88)
+
+___
+
+### projectRoot
+
+• **projectRoot**: `string`
+
+#### Defined in
+
+[src/types/debug-utils.ts:87](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L87)
+
+___
+
+### sources
+
+• **sources**: [`PersistSourceMapInput`](../classes/types_debug_utils.PersistSourceMapInput.md)[]
+
+#### Defined in
+
+[src/types/debug-utils.ts:86](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L86)
+
+___
+
+### withSources
+
+• `Optional` **withSources**: `boolean`
+
+#### Defined in
+
+[src/types/debug-utils.ts:89](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L89)
