+++ docs/code/interfaces/types_debug_utils.SimulateAndPersistResponseParams.md
@@ -1,0 +1,54 @@
+[@algorandfoundation/algokit-utils](../README.md) / [types/debug-utils](../modules/types_debug_utils.md) / SimulateAndPersistResponseParams
+
+# Interface: SimulateAndPersistResponseParams
+
+[types/debug-utils](../modules/types_debug_utils.md).SimulateAndPersistResponseParams
+
+## Table of contents
+
+### Properties
+
+- [algod](types_debug_utils.SimulateAndPersistResponseParams.md#algod)
+- [atc](types_debug_utils.SimulateAndPersistResponseParams.md#atc)
+- [bufferSizeMb](types_debug_utils.SimulateAndPersistResponseParams.md#buffersizemb)
+- [projectRoot](types_debug_utils.SimulateAndPersistResponseParams.md#projectroot)
+
+## Properties
+
+### algod
+
+• **algod**: `default`
+
+#### Defined in
+
+[src/types/debug-utils.ts:95](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L95)
+
+___
+
+### atc
+
+• **atc**: `AtomicTransactionComposer`
+
+#### Defined in
+
+[src/types/debug-utils.ts:93](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L93)
+
+___
+
+### bufferSizeMb
+
+• **bufferSizeMb**: `number`
+
+#### Defined in
+
+[src/types/debug-utils.ts:96](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L96)
+
+___
+
+### projectRoot
+
+• **projectRoot**: `string`
+
+#### Defined in
+
+[src/types/debug-utils.ts:94](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/debug-utils.ts#L94)
